[fix] API 명세서 반영 > Integer에서 Double

+++ backend/travel/src/main/java/com/travelrecommendation/travel/dto/TypeResult.java
@@ -11,5 +11,5 @@ import java.util.HashMap;
 public class TypeResult {
 
     private Integer type;
-    private HashMap<String,Integer> scores;
+    private HashMap<String,Double> scores;
 }
